Simplify RPC request hash table
Since the only operations are insert/delete, it is simpler to treat buckets with
`used` == 0 as deleted elements, forcing a linear walk over the whole table when
searching for inexistent keys during delete operations(responses).

Realistically this should not be an issue, as requests normally are replied
before new requests are sent.